Allow registering a Twin by passing an SP; refactor
Move Kotlin extension methods in a common lib.
