+++ lyo-services/webapp-twin/src/main/java/se/ericsson/cf/scott/sandbox/twin/services/AdaptorAdminService.java
@@ -17,6 +17,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import se.ericsson.cf.scott.sandbox.twin.TwinAdaptorHelper;
 import se.ericsson.cf.scott.sandbox.twin.TwinsServiceProviderInfo;
+import se.ericsson.cf.scott.sandbox.twin.clients.TwinRegistrationClient;
 import se.ericsson.cf.scott.sandbox.twin.servlet.TwinsServiceProvidersFactory;
 
 /**
@@ -27,36 +28,17 @@ import se.ericsson.cf.scott.sandbox.twin.servlet.TwinsServiceProvidersFactory;
 @Path("admin")
 public class AdaptorAdminService {
     private final static Logger log = LoggerFactory.getLogger(AdaptorAdminService.class);
-    
+
     @Context private HttpServletRequest httpServletRequest;
     @Context private HttpServletResponse httpServletResponse;
 
-    @POST
-    @Path("init")
-    @Produces({OslcMediaType.APPLICATION_RDF_XML, OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON, OslcMediaType.TEXT_TURTLE})
-    public Response getServiceProvider(@QueryParam("twinId") final String twinId)
-    {
-        log.info("Processing the init call");
+    final TwinRegistrationClient registrationClient = TwinAdaptorHelper.createTwinRegistrationClient();
 
-        final TwinsServiceProviderInfo twinInfo = new TwinsServiceProviderInfo("Fake Twin", "robot", twinId);
-        try {
-            // TODO Andrew@2019-01-28: refactor LyoD code to allow desc to be set here once
-            final ServiceProvider serviceProvider = TwinsServiceProvidersFactory.createTwinsServiceProvider(
-                twinInfo);
-            TwinAdaptorHelper.getServiceProviderRepository().addServiceProvider(serviceProvider);
-        } catch (OslcCoreApplicationException | URISyntaxException e) {
-            log.error("Can't init a dummy ServiceProvider", e);
-            return Response.status(Response.Status.BAD_REQUEST).build();
-        }
-
-        httpServletResponse.addHeader("Oslc-Core-Version","2.0");
-        return Response.noContent().build();
-    }
 
     @POST
     @Path("initRDF")
     @Consumes({OslcMediaType.APPLICATION_RDF_XML, OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON, OslcMediaType.TEXT_TURTLE})
-    @Produces({OslcMediaType.APPLICATION_RDF_XML, OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON, OslcMediaType.TEXT_TURTLE})
+//    @Produces({OslcMediaType.APPLICATION_RDF_XML, OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON, OslcMediaType.TEXT_TURTLE})
     public Response getServiceProvider(ServiceProvider sp)
     {
         log.info("Processing the init call (RDF input)");
@@ -67,6 +49,8 @@ public class AdaptorAdminService {
             final ServiceProvider serviceProvider = TwinsServiceProvidersFactory.createTwinsServiceProvider(
                 twinInfo);
             TwinAdaptorHelper.getServiceProviderRepository().addServiceProvider(serviceProvider);
+
+            registrationClient.registerTwin(serviceProvider);
         } catch (OslcCoreApplicationException | URISyntaxException e) {
             log.error("Can't init a dummy ServiceProvider", e);
             return Response.status(Response.Status.BAD_REQUEST).build();
